# openapi.yaml
openapi: 3.0.3
info:
  title: Integrated Travel Support Platform API
  description: REST API for unified trip planning, luggage storage discovery & booking, and last-mile transport booking.
  version: 1.1.0
servers:
  - url: http://localhost:4000/api/v1
    description: Local Development Server
paths:
...
  /payments:
    get:
      summary: List Traveler Payments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: related_type
          schema:
            type: string
            enum: [storage_booking, transport_booking]
      responses:
        '200':
          description: Paginated payment history
...
  /notifications:
    get:
      summary: List notifications for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paginated notifications list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
    post:
      summary: Send a data-driven notification (internal / admin dispatch)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - channel
                - template
                - payload
              properties:
                userId:
                  type: string
                  format: uuid
                channel:
                  type: string
                  enum: [email, sms, push]
                template:
                  type: string
                  enum: [BOOKING_CONFIRMED, BOOKING_CANCELLED, PICKUP_REMINDER_1H, PARTNER_STATUS_CHANGED]
                recipient:
                  type: string
                payload:
                  type: object
      responses:
        '201':
          description: Notification rendered, dispatched and recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
...
  /admin/partners:
    get:
      summary: Partner verification queue with status filtering and counts
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, verified, suspended, rejected]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated partner accounts with verification backlog counts
...
  /reviews:
    post:
      summary: Submit post-checkout review (1-5 stars + comment) tied to a completed booking
      description: Validates that the booking exists, belongs to the authenticated traveler, and has been completed (status checked_out for storage or completed for transport).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSubmissionInput'
      responses:
        '201':
          description: Review submitted successfully and audit event recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Review'
        '400':
          description: Validation error or booking is not checked out / completed
        '403':
          description: Forbidden - Traveler can only review their own bookings
        '409':
          description: Duplicate review - booking has already been reviewed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        channel:
          type: string
          enum: [email, sms, push]
        template:
          type: string
          enum: [BOOKING_CONFIRMED, BOOKING_CANCELLED, PICKUP_REMINDER_1H, PARTNER_STATUS_CHANGED]
        payload:
          type: object
        status:
          type: string
          enum: [pending, sent, delivered, failed]
        sent_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReviewSubmissionInput:
      type: object
      required:
        - bookingId
        - rating
      properties:
        bookingId:
          type: string
          example: storage-bk-101
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          maxLength: 1000
          nullable: true
          example: Seamless drop-off and pickup experience!
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
        relatedType:
          type: string
          enum: [storage_location, transport_option]
        relatedId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        author:
          type: object
          properties:
            id:
              type: string
            fullName:
              type: string
...